refactor(core): flatten repos/ into single storage.rs
The repos module (5 files, 293 lines) held the Storage trait and its
domain types — over-modularized for what it contained. Merge into one
storage.rs and rename repos:: to storage:: across the workspace.
