+++ crates/core/src/storage.rs
@@ -1,17 +1,18 @@
-//! Unified storage trait.
+//! Persistence trait and domain types.
 //!
-//! [`Storage`] is the single persistence backend for the runtime. It
-//! replaces the four sub-traits (`MemoryRepo`, `SkillRepo`, `SessionRepo`,
-//! `AgentRepo`) and their `Repos` composite with one flat interface.
-//! One implementation per backend — filesystem, in-memory, database.
+//! [`Storage`] is the unified persistence backend — one trait, one
+//! implementation per backend. Domain types ([`MemoryEntry`],
+//! [`SessionHandle`], [`Skill`]) live alongside the trait they serve.
 
 use crate::{
     AgentConfig, AgentId, ManifestConfig, NodeConfig,
     model::HistoryEntry,
-    repos::{MemoryEntry, SessionHandle, SessionSnapshot, SessionSummary, Skill},
     runtime::conversation::{ArchiveSegment, ConversationMeta, EventLine},
 };
-use anyhow::Result;
+use anyhow::{Result, bail};
+use std::collections::BTreeMap;
+
+// ── Storage trait ───────────────────────────────────────────────────
 
 /// Unified persistence backend.
 ///
@@ -123,4 +124,152 @@ pub trait Storage: Send + Sync + 'static {
 
     /// Create the initial config directory structure if it doesn't exist.
     fn scaffold(&self) -> Result<()>;
+}
+
+// ── Memory ──────────────────────────────────────────────────────────
+
+/// A single memory entry with YAML frontmatter metadata.
+#[derive(Debug, Clone)]
+pub struct MemoryEntry {
+    /// Human-readable name. Primary key for the repo (slugified for fs).
+    pub name: String,
+    /// One-line description used for relevance scoring.
+    pub description: String,
+    /// Entry body (markdown content).
+    pub content: String,
+}
+
+impl MemoryEntry {
+    /// Parse an entry from its frontmatter-based file content.
+    pub fn parse(raw: &str) -> Result<Self> {
+        let raw = raw.replace("\r\n", "\n");
+        let raw = raw.trim();
+        if !raw.starts_with("---") {
+            bail!("missing frontmatter opening ---");
+        }
+
+        let after_open = &raw[3..];
+        let Some(close_pos) = after_open.find("\n---") else {
+            bail!("missing frontmatter closing ---");
+        };
+
+        let frontmatter = &after_open[..close_pos];
+        let content = after_open[close_pos + 4..].trim().to_owned();
+
+        let mut name = None;
+        let mut description = None;
+
+        for line in frontmatter.lines() {
+            let line = line.trim();
+            if let Some(val) = line.strip_prefix("name:") {
+                name = Some(val.trim().to_owned());
+            } else if let Some(val) = line.strip_prefix("description:") {
+                description = Some(val.trim().to_owned());
+            }
+        }
+
+        let Some(name) = name else {
+            bail!("missing 'name' in frontmatter");
+        };
+        let description = description.unwrap_or_default();
+
+        Ok(Self {
+            name,
+            description,
+            content,
+        })
+    }
+
+    /// Serialize to the frontmatter file format.
+    pub fn serialize(&self) -> String {
+        let mut out = String::new();
+        out.push_str("---\n");
+        out.push_str(&format!("name: {}\n", self.name));
+        out.push_str(&format!("description: {}\n", self.description));
+        out.push_str("---\n\n");
+        out.push_str(&self.content);
+        out.push('\n');
+        out
+    }
+
+    /// Text for BM25 scoring — description + content concatenated.
+    pub fn search_text(&self) -> String {
+        format!("{} {}", self.description, self.content)
+    }
+}
+
+/// Convert a name to a filesystem-safe slug.
+pub fn slugify(name: &str) -> String {
+    let mut slug = String::with_capacity(name.len());
+    let mut prev_dash = true;
+
+    for ch in name.chars() {
+        if ch.is_alphanumeric() {
+            for lc in ch.to_lowercase() {
+                slug.push(lc);
+            }
+            prev_dash = false;
+        } else if !prev_dash {
+            slug.push('-');
+            prev_dash = true;
+        }
+    }
+
+    if slug.ends_with('-') {
+        slug.pop();
+    }
+
+    if slug.is_empty() {
+        slug.push_str("entry");
+    }
+
+    slug
+}
+
+// ── Sessions ────────────────────────────────────────────────────────
+
+/// Opaque handle identifying a persisted session. Created by the repo
+/// on `create`, returned by `find_latest`. Callers pass it back to
+/// append/load methods without interpreting the inner value.
+#[derive(Debug, Clone, PartialEq, Eq, Hash)]
+pub struct SessionHandle(String);
+
+impl SessionHandle {
+    /// Construct a handle from a repo-assigned identifier.
+    pub fn new(slug: impl Into<String>) -> Self {
+        Self(slug.into())
+    }
+
+    /// The raw identifier.
+    pub fn as_str(&self) -> &str {
+        &self.0
+    }
+}
+
+/// Snapshot returned by [`Storage::load_session`] — meta +
+/// working-context history, already replayed past the last compact
+/// marker.
+pub struct SessionSnapshot {
+    pub meta: ConversationMeta,
+    pub history: Vec<HistoryEntry>,
+}
+
+/// Summary returned by [`Storage::list_sessions`] for enumeration.
+pub struct SessionSummary {
+    pub handle: SessionHandle,
+    pub meta: ConversationMeta,
+}
+
+// ── Skills ──────────────────────────────────────────────────────────
+
+/// A named unit of agent behavior (agentskills.io format).
+#[derive(Debug, Clone)]
+pub struct Skill {
+    pub name: String,
+    pub description: String,
+    pub license: Option<String>,
+    pub compatibility: Option<String>,
+    pub metadata: BTreeMap<String, String>,
+    pub allowed_tools: Vec<String>,
+    pub body: String,
 }
